requirements.txt
setuptools
langchain
scikit-image
diffusers[torch]
torch
torchvision 
torchaudio
transformers
scipy
scikit-learn
pytest
invisible-watermark
safetensors
matplotlib
opencv-python
opencv-contrib-python
google-api-python-client
tiktoken
watchdog
openai
openai-whisper
altair==4.0
Flask
gunicorn
uuid
flask-swagger-ui
redis 
redisvl
prometheus-flask-exporter
pika
accelerate
safetensors
uvicorn
msgpack
grpcio 
grpcio-tools
pretty_errors
fastapi
datamodel-code-generator
ollama
rich
pynvml
python-multipart
apscheduler
minio
pydantic
#TTS @ git+https://github.com/coqui-ai/TTS@fa28f99f1508b5b5366539b2149963edcb80ba62
TTS==0.22.0
spacy==3.7.4
pymilvus
sentence-transformers 
pymupdf 
pytesseract 
pillow 
pydantic
faster-whisper
optimum
zmq